creada interfaz recursoDigital y renombrada recursoDigitalBase
