+++ src/biblioteca/gestorRecursos.java
@@ -5,17 +5,17 @@ import java.util.List;
 
 //Con gestor recursos podemos recorrer los recursos en una lista y agregarlos a ella
 public class gestorRecursos {
-    private List<recursoDigital> recursos;
+    private List<recursoDigitalBase> recursos;
 
     public gestorRecursos() {
         recursos = new ArrayList<>();
     }
 
-    public void agregarRecurso(recursoDigital recurso) {
+    public void agregarRecurso(recursoDigitalBase recurso) {
         recursos.add(recurso);
     }
 
-    public List<recursoDigital> getRecursos() {
+    public List<recursoDigitalBase> getRecursos() {
         return recursos;
     }
 }
